More charms.docker goodness
refactor the one-off subprocess calls with a docker-compose based launch.
This templatizes both the agent (all machines), and the manager(on the leader)
and runs on port 2376, this is prelimiary as the TLS terminated docker socket
is *supposed* to reside on this port, but we'll be placing swarm there instead.